YQL-19747: Split tests by component
