+++ yql/essentials/utils/cache/ut/ya.make
@@ -1,6 +1,8 @@
 UNITTEST_FOR(yql/essentials/utils/cache)
 
 SRCS(
+    managed_cache_maintenance_ut.cpp
+    managed_cache_storage_ut.cpp
     managed_cache_ut.cpp
 )
